phase_2/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
better-exceptions==0.3.3
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
cryptography==41.0.8
uvicorn==0.24.0.post1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
PyJWT==2.8.0
jwcrypto==1.5.6